+++ tests/tests-main.cpp
@@ -472,13 +472,14 @@ TEST_CASE( "Example 3: multi-env w/ covars" ){
 	}
 }
 
-TEST_CASE( "Example 4: multi-env w/ covars" ){
+TEST_CASE( "Example 4: multi-env + mog + covars + emp_bayes" ){
 	parameters p;
 
 	SECTION("Ex4. No filters applied, high mem mode"){
-		char* argv[] = { (char*) "bin/bgen_prog", (char*) "--mode_vb", (char*) "--high_mem",
+		char* argv[] = { (char*) "bin/bgen_prog", (char*) "--mode_vb", (char*) "--low_mem",
 						 (char*) "--use_vb_on_covars", (char*) "--mode_empirical_bayes",
-						 (char*) "--vb_iter_max", (char*) "30",
+						 (char*) "--effects_prior_mog",
+						 (char*) "--vb_iter_max", (char*) "10",
 						 (char*) "--environment", (char*) "data/io_test/n50_p100_env.txt",
 						 (char*) "--bgen", (char*) "data/io_test/n50_p100.bgen",
 						 (char*) "--out", (char*) "data/io_test/fake_env.out",
@@ -518,7 +519,7 @@ TEST_CASE( "Example 4: multi-env w/ covars" ){
 			CHECK(VB.n_effects == 2);
 			CHECK(VB.vp_init.muw(0) == 0.25);
 			CHECK(VB.p.init_weights_with_snpwise_scan == false);
-			CHECK(VB.dXtEEX(0, 0) == Approx(38.9390135703));
+			CHECK(VB.dXtEEX(0, 0) == Approx(38.9963604264));
 		}
 
 		std::vector< VbTracker > trackers(VB.hyps_grid.rows());
@@ -540,27 +541,52 @@ TEST_CASE( "Example 4: multi-env w/ covars" ){
 			VariationalParameters& vp = all_vp[0];
 			Hyps& hyps = all_hyps[0];
 
-			CHECK(vp.alpha_beta(0)          == Approx(0.1339235799));
-			CHECK(vp.alpha_beta(1)          == Approx(0.1415361555));
-			CHECK(vp.alpha_beta(63)         == Approx(0.1724736345));
-			CHECK(vp.muw(0, 0)              == Approx(0.1127445891));
-			CHECK(hyps.sigma                == Approx(0.6953636046));
-			CHECK(hyps.lambda[0]            == Approx(0.1698152775));
+			CHECK(vp.alpha_beta(0)            == Approx(0.1339032153));
+			CHECK(vp.alpha_beta(1)            == Approx(0.1413799597));
+			CHECK(vp.alpha_beta(63)           == Approx(0.1722999226));
+			CHECK(vp.muw(0, 0)              == Approx(0.1129510095));
+
+			CHECK(hyps.sigma                == Approx(0.6956395607));
+			CHECK(hyps.lambda[0]            == Approx(0.1697910485));
 			CHECK(hyps.lambda[1]            == Approx(0.1350333198));
-			CHECK(hyps.slab_relative_var[0] == Approx(0.0081893295));
-			CHECK(hyps.slab_relative_var[1] == Approx(0.005117116));
+			CHECK(hyps.slab_relative_var[0] == Approx(0.0081833922));
+			CHECK(hyps.slab_relative_var[1] == Approx(0.0051150827));
 
 			VB.updateAllParams(1, round_index, all_vp, all_hyps, logw_prev, trackers, logw_updates);
 
-			CHECK(vp.alpha_beta(0)          == Approx(0.1475535664));
-			CHECK(vp.alpha_beta(1)          == Approx(0.1539872875));
-			CHECK(vp.alpha_beta(63)         == Approx(0.3139615442));
-			CHECK(vp.muw(0, 0)              == Approx(0.062497288));
-			CHECK(hyps.sigma                == Approx(0.6027825451));
-			CHECK(hyps.lambda[0]            == Approx(0.1994659195));
-			CHECK(hyps.lambda[1]            == Approx(0.1169821203));
-			CHECK(hyps.slab_relative_var[0] == Approx(0.0125774267));
-			CHECK(hyps.slab_relative_var[1] == Approx(0.0042557827));
+			CHECK(vp.alpha_beta(0)            == Approx(0.1475445081));
+			CHECK(vp.muw(0, 0)              == Approx(0.0627874101));
+			CHECK(vp.alpha_gam(63)           == Approx(0.1183555819));
+			CHECK(vp.mu1_gam(63)              == Approx(0.0013071544));
+			CHECK(vp.s1_gam_sq(63)            == Approx(0.0026294351));
+
+			CHECK(hyps.sigma                == Approx(0.6031329457));
+			CHECK(hyps.lambda[0]            == Approx(0.1994253645));
+			CHECK(hyps.lambda[1]            == Approx(0.1169875913));
+			CHECK(hyps.slab_relative_var[0] == Approx(0.0125648575));
+			CHECK(hyps.slab_relative_var[1] == Approx(0.0042537538));
+			CHECK(hyps.s_x[0]               == Approx(64.0));
+			CHECK(hyps.s_x[1]               == Approx(0.2658034479));
+			CHECK(hyps.pve[1]               == Approx(0.0001139822));
+			CHECK(hyps.pve_large[1]         == Approx(0.000113981));
+
+			VB.updateAllParams(2, round_index, all_vp, all_hyps, logw_prev, trackers, logw_updates);
+
+			CHECK(vp.alpha_beta(63)           == Approx(0.2849674292));
+			CHECK(vp.muw(0, 0)              == Approx(0.0385493631));
+			CHECK(vp.alpha_gam(63)           == Approx(0.1035741368));
+			CHECK(vp.mu1_gam(63)              == Approx(-0.0008068742));
+			CHECK(vp.s1_gam_sq(63)            == Approx(0.0019506006));
+
+			CHECK(hyps.sigma                == Approx(0.5604876755));
+			CHECK(hyps.lambda[0]            == Approx(0.2187750882));
+			CHECK(hyps.lambda[1]            == Approx(0.1024356641));
+			CHECK(hyps.slab_relative_var[0] == Approx(0.0158099281));
+			CHECK(hyps.slab_relative_var[1] == Approx(0.0033998137));
+			CHECK(hyps.s_x[0]               == Approx(64.0));
+			CHECK(hyps.s_x[1]               == Approx(0.1035936049));
+			CHECK(hyps.pve[1]               == Approx(0.0000295386));
+			CHECK(hyps.pve_large[1]         == Approx(0.0000295381));
 		}
 	}
 }